Refactor: extract duplicate code to TelemetryCommandHelper
Extract common utility methods shared across telemetry commands into
a new TelemetryCommandHelper static class to improve code consistency
and reduce duplication:
- GetDashboardConnection: shared Dashboard connection logic
- GetTextFromResult: MCP result text extraction
- ExtractJsonFromResult: JSON extraction from MCP responses
- EscapeJsonString: JSON string escaping

This reduces ~200 lines of duplicate code across the four telemetry
commands (traces, logs, fields, metrics).
